\name{cuteMatrix-package}
\alias{cuteMatrix-package}
\alias{cuteMatrix}
\docType{package}
\title{
\packageTitle{cuteMatrix}
}
\description{
\packageDescription{cuteMatrix}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{cuteMatrix}
\packageIndices{cuteMatrix}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{cuteMatrix}

Maintainer: \packageMaintainer{cuteMatrix}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
