% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_op.R
\name{mat_op}
\alias{mat_op}
\title{mat_op}
\usage{
mat_op(mat.list, kind.of.operation = "+")
}
\arguments{
\item{mat.list}{List of matrices.}

\item{kind.of.operation}{Either "+" (by case addition), "-" (by case subtraction) or "*" (by case multiplication).}
}
\value{
The assembled matrix, with row and/or column names only if all the matrices have identical row/column names.
}
\description{
assemble several matrices of same dimensions by performing by case operation. For instance, with kind.of.operation = "+", the function will add the value of all the case 1 (row1 & column1) of the matrices and put it in the case 1 of a new matrix M, according to the formula

\deqn{c_{ij} = \sum_{k=1}^{k=z}c_{ijk}}
c: case

i: row number

j: column number

k: matrix number

z: number of matrices
}
\details{
REQUIRED PACKAGES

cuteDev

cuteTool

REQUIRED FUNCTIONS FROM THE cute PACKAGE

arg_check()

comp_2d()
}
\examples{
mat1 = matrix(c(1,1,1,2,1,5,9,8), ncol = 2) ; 
mat2 = matrix(c(1,1,1,2,1,5,9,NA), ncol = 2) ; 
mat_op(mat.list = list(mat1, mat2), kind.of.operation = "+")

mat1 = matrix(c(1,1,1,2,1,5,9,8), ncol = 2, dimnames = list(LETTERS[1:4], letters[1:2])) ; 
mat2 = matrix(c(1,1,1,2,1,5,9,NA), ncol = 2, dimnames = list(LETTERS[1:4], letters[1:2])) ; 
mat_op(mat.list = list(mat1, mat2), kind.of.operation = "*")

mat1 = matrix(c(1,1,1,2,1,5,9,8), ncol = 2, dimnames = list(LETTERS[1:4], c(NA, NA))) ; 
mat2 = matrix(c(1,1,1,2,1,5,9,NA), ncol = 2, dimnames = list(LETTERS[1:4], letters[1:2])) ; 
mat_op(mat.list = list(mat1, mat2), kind.of.operation = "-")

mat1 = matrix(c(1,1,1,2,1,5,9,8), ncol = 2, dimnames = list(c("A1", "A2", "A3", "A4"), letters[1:2])) ; 
mat2 = matrix(c(1,1,1,2,1,5,9,NA), ncol = 2, dimnames = list(LETTERS[1:4], letters[1:2])) ; 
mat3 = matrix(c(1,1,1,2,1,5,9,NA), ncol = 2, dimnames = list(LETTERS[1:4], letters[1:2])) ; 
mat_op(mat.list = list(mat1, mat2, mat3), kind.of.operation = "+")
}
