% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_rotate.R
\name{mat_rotate}
\alias{mat_rotate}
\title{mat_rotate}
\usage{
mat_rotate(data)
}
\arguments{
\item{data}{Matrix (matrix class).}
}
\value{
The modified matrix.
}
\description{
90° clockwise matrix rotation.

Applied twice, the function provide the mirror matrix, according to vertical and horizontal symmetry.
}
\examples{
obs <- matrix(1:10, ncol = 1) ; obs ; mat_rotate(obs)

obs <- matrix(LETTERS[1:10], ncol = 5) ; obs ; mat_rotate(obs)
}
