% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_inv.R
\name{mat_inv}
\alias{mat_inv}
\title{mat_inv}
\usage{
mat_inv(mat, safer_check = TRUE)
}
\arguments{
\item{mat}{A square numeric matrix without NULL, NA, Inf or single case (dimension 1, 1) of 0.}

\item{safer_check}{Single logical value. Perform some "safer" checks (see https://github.com/safer-r)? If TRUE, checkings are performed before main code running: 1) R classical operators (like "<-") not overwritten by another package because of the R scope and 2) required functions and related packages effectively present in local R lybraries. Set to FALSE if this fonction is used inside another "safer" function to avoid pointless multiple checkings.}
}
\value{
The inversed matrix.
}
\description{
Return the inverse of a square matrix when solve() cannot.
}
\examples{
mat1 = matrix(c(1,1,1,2,1,5,9,8,9), ncol = 3) ; mat_inv(mat = mat1) # use solve()

mat1 = matrix(2, ncol = 1) ; mat_inv(mat = mat1)
}
\seealso{
\code{\link[MASS]{ginv}}, \code{\link[pracma]{pinv}} which can calculates the Moore-Penrose generalized inverse of a matrix X.
}
\author{
Gael Millot <gael.millot@pasteur.fr>

Yushi Han <yushi.han2000@gmail.com>

Haiding Wang <wanghaiding442@gmail.com>
}
