% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_inv.R
\name{mat_inv}
\alias{mat_inv}
\title{mat_inv}
\usage{
mat_inv(mat)
}
\arguments{
\item{mat}{A square numeric matrix without NULL, NA, Inf or single case (dimension 1, 1) of 0.}
}
\value{
The inversed matrix.
}
\description{
Return the inverse of a square matrix when solve() cannot.
}
\examples{
mat1 = matrix(c(1,1,1,2,1,5,9,8,9), ncol = 3) ; mat_inv(mat = mat1) # use solve()

mat1 = matrix(2, ncol = 1) ; mat_inv(mat = mat1)
}
