% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_rotate.R
\name{mat_rotate}
\alias{mat_rotate}
\title{mat_rotate}
\usage{
mat_rotate(data, safer_check = TRUE)
}
\arguments{
\item{data}{Matrix (matrix class).}

\item{safer_check}{Single logical value. Perform some "safer" checks (see https://github.com/safer-r)? If TRUE, checkings are performed before main code running: 1) R classical operators (like "<-") not overwritten by another package because of the R scope and 2) required functions and related packages effectively present in local R lybraries. Set to FALSE if this fonction is used inside another "safer" function to avoid pointless multiple checkings.}
}
\value{
The modified matrix.
}
\description{
90° clockwise matrix rotation.

Applied twice, the function provide the mirror matrix, according to vertical and horizontal symmetry.
}
\examples{
obs <- matrix(1:10, ncol = 1) ; obs ; mat_rotate(obs)

obs <- matrix(LETTERS[1:10], ncol = 5) ; obs ; mat_rotate(obs)
}
\author{
Gael Millot <gael.millot@pasteur.fr>

Yushi Han <yushi.han2000@gmail.com>

Haiding Wang <wanghaiding442@gmail.com>
}
